requirements.txt
conda-build
gitpython
conda-build-all
